Change flow of entire game
Matching -> moved to stage 1
Memorising -> moved to stage 2
Answering -> moved to stage 3
